+++ client/src/App.jsx
@@ -1,64 +1,59 @@
 import React, { useState, useContext, useEffect } from "react";
-import { ThemeProvider } from "styled-components/macro";
 import { Switch, Route } from "react-router-dom";
-import axios from "axios";
-import Row from "react-bootstrap/Row";
-import Col from "react-bootstrap/Col";
+import { ThemeProvider } from "styled-components/macro";
+import { COLOR, GlobalStyle, ResetStyle } from "./components/GlobalStyle";
+import { AuthContext } from "./contexts/AuthContext";
+import UserService from "./services/UserService";
+import { RestaurantsContext } from "./contexts/RestaurantsContext";
+import RestaurantsService from "./services/RestaurantsService";
 import Login from "./views/Login";
 import Place from "./views/Place";
 import FavPlaces from "./views/FavPlaces";
 import Home from "./views/Home";
 import Loader from "./components/Loader";
 import Navigation from "./components/Navigation";
-import { COLOR, GlobalStyle, ResetStyle } from "./components/GlobalStyle";
-import { AuthContext } from "./contexts/AuthContext";
-import RestaurantsService from "./services/RestaurantsService";
-import { RestaurantsContext } from "./contexts/RestaurantsContext";
 
 function App() {
   const { auth, setAuth } = useContext(AuthContext);
-  const { restaurants, setRestaurants } = useContext(RestaurantsContext);
+  const { setRestaurants } = useContext(RestaurantsContext);
   const [showSidebar, setShowSidebar] = useState(false);
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {}, [setRestaurants]);
 
   useEffect(() => {
-    //axios.get used GET request to fetch user data from MongoDB
-    axios
-      .get("/user", { withCredentials: true })
+    // Get user data
+    UserService.getUser()
       .then((response) => {
         setAuth(response.data);
         setShowSidebar(true);
 
+        // Get restaurants data
         RestaurantsService.getAll()
           .then((res) => {
             setRestaurants(res.data);
             setLoading(false);
           })
           .catch((error) => {
+            console.log(error.response.data.message);
             setLoading(false);
           });
       })
       .catch((error) => {
+        console.log(error.response.data.message);
         setLoading(false);
       });
-  }, []);
+  }, [setAuth, setRestaurants]);
 
   return (
     <>
       <ThemeProvider theme={COLOR.light}>
         <ResetStyle />
         <GlobalStyle />
-
+        <Navigation showSidebar={showSidebar} setShowSidebar={setShowSidebar} />
         <Switch>
           {!loading && auth && (
             <>
-              <Navigation
-                showSidebar={showSidebar}
-                setShowSidebar={setShowSidebar}
-                auth={auth}
-              />
               <Route exact path="/">
                 <Home />
               </Route>
@@ -72,12 +67,9 @@ function App() {
           )}
           {loading && <Loader />}
           {!auth && (
-            <>
-              <Navigation />
-              <Route exact path="/">
-                <Login setShowSidebar={setShowSidebar} />
-              </Route>
-            </>
+            <Route exact path="/">
+              <Login setShowSidebar={setShowSidebar} />
+            </Route>
           )}
         </Switch>
       </ThemeProvider>
